PF-1829 redox gem implement media push apis (#3)
* Implement file upload; fix re-auth bug

Redox has a separate API for uploading files with arbitrary content.
It does not use any of the Redox models. Instead it is a standard
POST of a multipart form to a fixed endpoint.

Redox::FileUpload implements support for this file upload.

This entailed updating VCR related code to handle base64 data.

This also fixes a bug where Redox::Source would reauthenticate
for every request. The bug was adding integer values to DateTimes.
This adds days, not seconds. Updated tests to catch and fixed this bug.

* Implement Media.New

* Remove invalid query model properties

* Add VCR_MODE for more convenient recording